pyproject.toml
[project]
name = "dlpctl"
version = "0.1.0"
description = "This application is used to control a Vialux DLP for lab usage"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["alp4lib>=1.0.2", "numpy>=2.2.6"]
